Redesign Choclo's API (#12)
Create a new point submodule that hosts kernel functions and forward
modelling functions for point sources. Add test functions for the new forward
modelling functions and edit the ones for the kernels. Add a constants.py
file for hosting the universal constants needed to perform the forward
modelling. Change the parameters of the distance functions: they now take the
coordinates of the points as floats instead of grouped in tuples.