Add magnetic fields of a dipole (#15)
Add forward modelling functions for the magnetic field of a dipole:
magnetic_e, magnetic_n, magnetic_u. Add a magnetic_field that returns
the three components of the magnetic field of a dipole in a faster way than
running each one of the previous functions. Add symmetry tests for each
component, test if magnetic_field generates the same values as each one of
the component functions and test if the divergence of the magnetic field is
zero. Add the new functions to the API reference.