Fix bug on non-diagonal tensor components of prisms (#27)
Fix a bug on the non-diagonal tensor components of the prism due to a wrong
evaluation of the logarithmic function. Add tests that catch the bug on every
non-diagonal tensor component. Correctly evaluate the log function present in
the non-diagonal tensor components when the shifted coordinate `x` of the
vertex is negative and it's equal to the negative of the distance `r` between
the vertex and the computation point. In that case the `log(x + r)` can be
replaced by `-log(2|x|)`. Also, implement a modified version of the log
function proposed by Fukushima (2020) for cases where the shifted coordinate of
the node is negative but different than the negative of the distance: improves
the accuracy when evaluating the log function on very small values. Now the
`_safe_log` function takes two arguments instead of one: the shifted coordinate
`x` and the distance `r`. Minor improvements to the docstrings.