Implement singular points for diagonal tensor components
Return nan when computing the diagonal tensor components on singular
points (vertices or edges perpendicular to the direction of the tensor
component). When computing on the greatest face (eastern,northern or top
face) make the respective diagonal component to return the limit
approaching from the outside. There's no need to change the
implementation, just add a term of 4 pi to the result.