Drop support for Python 3.6 (#29)
Drop support for Python 3.6 and revert changes made in #28. After almost a year
that Python 3.6 reached its end of life, it makes sense to drop the support for
it in Choclo. Moreover, all the other libraries in Fatiando require Python
versions greater or equal to 3.7.

References: https://devguide.python.org/versions/