Start drafting singular points for magnetic forwards
Start creating special cases for singular points on the magnetic forward
modelling for prisms. Returns nans when observation point falls in prism
vertices and edges. Solves the issue with the limit on east, north and
top faces. Lacks of tests and decide what to do with internal points
(where the magnetic field integral is not valid anymore).