Ditch kernel functions for point masses (#42)
The kernel functions for gravity fields of point sources were just one
line functions with a lot of docstrings and their own tests. We are not
using them outsides the actual gravity forward modelling functions for
point masses, so it might be better to remove them. We can always
restore them if we find a use case for them.

**Relevant issues/PRs:**

Fixes #38