Prism forward models return nan on singular points (#30)
Gravity and magnetic forward modelling functions for prisms return
`np.nan` when the observation points falls in a singular point. These
singular points vary for the field that is being computed. For the
gravity tensor components, these are all the vertices and edges
perpendicular to the direction(s) of the tensor component. For the
magnetic fields, these are all the vertices, all the edges and all the
interior points. This PR also fixes the values of the fields on the
faces: diagonal gravity tensor components and magnetic field components
on prism faces return the limit of the field approaching from outside.
To achieve that, a 4π constant must be added when computing on the east,
north or top faces. Add tests for these new features and include
explanations in the docstrings.

**Relevant issues**

Fixes #36