Take only scalar inputs instead of array inputs  (#50)
Choclo functions shouldn't take array inputs because they may require a
user to convert their prism, dipole moment, or magnetization format into
what Choclo expects with the possible result of reallocating memory or
having to iterate over slices in non-optimal ways. Since these functions
are meant to be wrapped by other JIT-compiled functions that the user
creates, they should be as flexible as possible with their inputs. This
means converting the `dipole_moment`, `prism`, and `magnetization`
arguments into individual components instead of arrays. The code is now
a bit longer because of this but there is no other down side. Had to
update some linting rules to account for the longer code and larger
number of arguments that this entails.