Update the versions of sphinx and its plugins (#64)
Updates of the theme required minor tweaks to the configuration, mostly
removing the logo from the left side bar. Had to remove the caching of
conda packages from the CI because conda fails to find packages when
it's enabled. Don't know why but remove it for now since I found no way
around it.