Fix sign of third-order kernels
Update the signs in the documentation and in the implementation of each
kernel. Also make the kernels to return zero on singular points (not on
vertices) so when computing on lines shared with the edges are not falling
in division by zero errors and the limits for those cases are well
resolved.