Fix bug on `safe_log` and solves discontinuous magnetic fields (#100)
Fix how `_safe_log` checks if `r == |x|`. Add the other two shifted
coordinates (`y` and `z`) as arguments for `_safe_log` so we can use
them to assess that the observation point is inline with a pair of
vertices or not. Make the check by evaluating if `y == 0 and z == 0`
instead of `-x == r` (with `x < 0`), since `-x` and `r` could be the
same value (up to machine precision) while `y` and `z` are non-zero
(specially when $|x| \gg |y|$ and/or $|x| \gg |z|$). Update calls to
`_safe_log`, update its docstrings, and update docstrings of kernels
that mentioned the safe log function. Add tests that catches the bug,
and that pass with the bugfix.