Add magnetic gradiometry forwards for prisms (#97)
Add forward modelling functions for the magnetic gradiometry components
due to rectangular prisms: spatial derivatives of the magnetic field
components. Include third-order kernel functions that are needed to
compute the derivatives of the magnetic field components. Create private
function to evaluate the kernels on the vertices of the prism provided
its magnetization so we can reduce the amount of repeated code in the
magnetic forward modelling functions. Add tests for the new functions.