Use zizmor to lint GitHub Actions (#114)
Make use of `zizmor` to lint GitHub Actions workflows and catch
potential security issues. Add `zizmor` to the `environment.yml` and to
the `requirements-style.txt`. Add a new `check-actions` target in the
`Makefile` that runs `zizmor` on every workflow. Add a new workflow for
running `zizmor` on PRs and on pushes to `main`. Apply some minor
changes to the script that deploys docs to GitHub Pages after
suggestions obtained by running `shellcheck` on it.