Fix deployment of docs (#122)
Persist credentials after checking out the `gh-pages` in the
documentation workflow: we need those credentials to be able to push
changes to the `gh-pages` branch. Use `set -e` in the deployment script
so the workflow is marked as failed if any command exists with non-zero
status. Remove Ensaio caching from workflow, since Choclo doesn't use
Ensaio for building docs, so we don't need to cache Ensaio files.


**Relevant issues/PRs:**

Solve what #121 was trying to solve.