Fix bug in third order prism kernels (#118)
Fix bug related to numerical instabilities on third order kernels
`kernel_iij`, where `i, j in {x, y, z}`. Make use of the alternative
definition of the kernels. These two different definition are not
mathematically equal, but their difference depends solely on two of the
three coordinates, therefore when evaluating the kernels on the prism
boundaries that extra term gets cancelled out. This property makes these
two different definitions equivalent in the purpose of forward modelling
prism fields. Add tests that catch the bug and ensure that the new
implementation works as expected. Update the docstrings.