Move to a src layout
Moving the package code to `src` and tests to `test` makes it easier to
make sure we're testing the installed version (don't need to make a temp
dir anymore) and avoids distributing the testing code with the package,
which was not very useful and reduces the size of the downloads. This
simplifies the Makefile as well.