Remove level 2 headings from API docs (#135)
The API docs had level 2 headings that were "Kernel and forward modeling
functions" and "Utilities". This is what shows up in the right
navigation bar of the docs. So it's not easy to jump to the prism
functions, for example. Remove the current level 2 headings (which
promotes the level 3 headings to level 2) for better navigation. Now we
can click to go to the prism docs as soon as the page is loaded. It also
provides a better overview of what the contents are. Also change the
module description in `src/choclo/__init__.py` so that the automodule
call doesn't add a description of Choclo itself but says that this is
the Choclo API. Add `utils` and `constants` as imports in `__init__.py`
for consistency with other modules.